Add split js files and import to index.js
